--- a/tsdb/index/postings.go
+++ b/tsdb/index/postings.go
@@ -26,6 +26,7 @@ import (
 	"sync"
 
 	"github.com/bboreham/go-loser"
+	"github.com/dolthub/swiss"
 
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/storage"
@@ -55,14 +56,14 @@ var ensureOrderBatchPool = sync.Pool{
 // unordered batch fills on startup.
 type MemPostings struct {
 	mtx     sync.RWMutex
-	m       map[string]map[string][]storage.SeriesRef
+	m       map[string]*swiss.Map[string, []storage.SeriesRef]
 	ordered bool
 }
 
 // NewMemPostings returns a memPostings that's ready for reads and writes.
 func NewMemPostings() *MemPostings {
 	return &MemPostings{
-		m:       make(map[string]map[string][]storage.SeriesRef, 512),
+		m:       make(map[string]*swiss.Map[string, []storage.SeriesRef], 512),
 		ordered: true,
 	}
 }
@@ -71,7 +72,7 @@ func NewMemPostings() *MemPostings {
 // until EnsureOrder() was called once.
 func NewUnorderedMemPostings() *MemPostings {
 	return &MemPostings{
-		m:       make(map[string]map[string][]storage.SeriesRef, 512),
+		m:       make(map[string]*swiss.Map[string, []storage.SeriesRef], 512),
 		ordered: false,
 	}
 }
@@ -84,9 +85,10 @@ func (p *MemPostings) Symbols() StringIter {
 	symbols := make(map[string]struct{}, 512)
 	for n, e := range p.m {
 		symbols[n] = struct{}{}
-		for v := range e {
+		e.Iter(func(v string, _ []storage.SeriesRef) bool {
 			symbols[v] = struct{}{}
-		}
+			return false
+		})
 	}
 	p.mtx.RUnlock()
 
@@ -105,9 +107,10 @@ func (p *MemPostings) SortedKeys() []labels.Label {
 	keys := make([]labels.Label, 0, len(p.m))
 
 	for n, e := range p.m {
-		for v := range e {
+		e.Iter(func(v string, _ []storage.SeriesRef) bool {
 			keys = append(keys, labels.Label{Name: n, Value: v})
-		}
+			return false
+		})
 	}
 	p.mtx.RUnlock()
 
@@ -146,10 +149,16 @@ func (p *MemPostings) LabelValues(_ context.Context, name string) []string {
 	p.mtx.RLock()
 	defer p.mtx.RUnlock()
 
-	values := make([]string, 0, len(p.m[name]))
-	for v := range p.m[name] {
-		values = append(values, v)
+	e := p.m[name]
+	if e == nil || e.Count() == 0 {
+		return nil
 	}
+
+	values := make([]string, 0, e.Count())
+	e.Iter(func(v string, _ []storage.SeriesRef) bool {
+		values = append(values, v)
+		return false
+	})
 	return values
 }
 
@@ -182,17 +191,18 @@ func (p *MemPostings) Stats(label string, limit int) *PostingsStats {
 		if n == "" {
 			continue
 		}
-		labels.push(Stat{Name: n, Count: uint64(len(e))})
-		numLabelPairs += len(e)
+		labels.push(Stat{Name: n, Count: uint64(e.Count())})
+		numLabelPairs += e.Count()
 		size = 0
-		for name, values := range e {
+		e.Iter(func(name string, values []storage.SeriesRef) bool {
 			if n == label {
 				metrics.push(Stat{Name: name, Count: uint64(len(values))})
 			}
 			seriesCnt := uint64(len(values))
 			labelValuePairs.push(Stat{Name: n + "=" + name, Count: seriesCnt})
 			size += uint64(len(name)) * seriesCnt
-		}
+			return false
+		})
 		labelValueLength.push(Stat{Name: n, Count: size})
 	}
 
@@ -213,7 +223,7 @@ func (p *MemPostings) Get(name, value string) Postings {
 	p.mtx.RLock()
 	l := p.m[name]
 	if l != nil {
-		lp = l[value]
+		lp, _ = l.Get(value)
 	}
 	p.mtx.RUnlock()
 
@@ -266,14 +276,16 @@ func (p *MemPostings) EnsureOrder(numberOfConcurrentProcesses int) {
 
 	nextJob := ensureOrderBatchPool.Get().(*[][]storage.SeriesRef)
 	for _, e := range p.m {
-		for _, l := range e {
+		e.Iter(func(_ string, l []storage.SeriesRef) bool {
 			*nextJob = append(*nextJob, l)
 
 			if len(*nextJob) >= ensureOrderBatchSize {
 				workc <- nextJob
 				nextJob = ensureOrderBatchPool.Get().(*[][]storage.SeriesRef)
 			}
-		}
+
+			return false
+		})
 	}
 
 	// If the last job was partially filled, we need to push it to workers too.
@@ -294,7 +306,11 @@ func (p *MemPostings) Delete(deleted map[storage.SeriesRef]struct{}, affected ma
 	defer p.mtx.Unlock()
 
 	process := func(l labels.Label) {
-		orig := p.m[l.Name][l.Value]
+		nm := p.m[l.Name]
+		var orig []storage.SeriesRef
+		if nm != nil {
+			orig, _ = nm.Get(l.Value)
+		}
 		repl := make([]storage.SeriesRef, 0, len(orig))
 		for _, id := range orig {
 			if _, ok := deleted[id]; !ok {
@@ -302,11 +318,13 @@ func (p *MemPostings) Delete(deleted map[storage.SeriesRef]struct{}, affected ma
 			}
 		}
 		if len(repl) > 0 {
-			p.m[l.Name][l.Value] = repl
+			nm.Put(l.Value, repl)
 		} else {
-			delete(p.m[l.Name], l.Value)
+			if nm != nil {
+				nm.Delete(l.Value)
+			}
 			// Delete the key if we removed all values.
-			if len(p.m[l.Name]) == 0 {
+			if nm == nil || nm.Count() == 0 {
 				delete(p.m, l.Name)
 			}
 		}
@@ -324,10 +342,13 @@ func (p *MemPostings) Iter(f func(labels.Label, Postings) error) error {
 	defer p.mtx.RUnlock()
 
 	for n, e := range p.m {
-		for v, p := range e {
-			if err := f(labels.Label{Name: n, Value: v}, newListPostings(p...)); err != nil {
-				return err
-			}
+		var err error
+		e.Iter(func(v string, p []storage.SeriesRef) bool {
+			err = f(labels.Label{Name: n, Value: v}, newListPostings(p...))
+			return err != nil
+		})
+		if err != nil {
+			return err
 		}
 	}
 	return nil
@@ -348,11 +369,12 @@ func (p *MemPostings) Add(id storage.SeriesRef, lset labels.Labels) {
 func (p *MemPostings) addFor(id storage.SeriesRef, l labels.Label) {
 	nm, ok := p.m[l.Name]
 	if !ok {
-		nm = map[string][]storage.SeriesRef{}
+		nm = swiss.NewMap[string, []storage.SeriesRef](0)
 		p.m[l.Name] = nm
 	}
-	list := append(nm[l.Value], id)
-	nm[l.Value] = list
+	list, _ := nm.Get(l.Value)
+	list = append(list, id)
+	nm.Put(l.Value, list)
 
 	if !p.ordered {
 		return
@@ -401,7 +423,7 @@ func (p *MemPostings) PostingsForLabelMatching(ctx context.Context, name string,
 	p.mtx.RLock()
 	e := p.m[name]
 	for _, v := range vals {
-		if refs, ok := e[v]; ok {
+		if refs, ok := e.Get(v); ok {
 			// Some of the values may have been garbage-collected in the meantime this is fine, we'll just skip them.
 			// If we didn't let the mutex go, we'd have these postings here, but they would be pointing nowhere
 			// because there would be a `MemPostings.Delete()` call waiting for the lock to delete these labels,
@@ -422,16 +444,17 @@ func (p *MemPostings) labelValues(name string) []string {
 	defer p.mtx.RUnlock()
 
 	e := p.m[name]
-	if len(e) == 0 {
+	if e == nil || e.Count() == 0 {
 		return nil
 	}
 
-	vals := make([]string, 0, len(e))
-	for v, srs := range e {
+	vals := make([]string, 0, e.Count())
+	e.Iter(func(v string, srs []storage.SeriesRef) bool {
 		if len(srs) > 0 {
 			vals = append(vals, v)
 		}
-	}
+		return false
+	})
 
 	return vals
 }