--- a/tsdb/index/postings_test.go
+++ b/tsdb/index/postings_test.go
@@ -26,6 +26,7 @@ import (
 	"sync"
 	"testing"
 
+	"github.com/dolthub/swiss"
 	"github.com/grafana/regexp"
 	"github.com/stretchr/testify/require"
 
@@ -36,17 +37,18 @@ import (
 
 func TestMemPostings_addFor(t *testing.T) {
 	p := NewMemPostings()
-	p.m[allPostingsKey.Name] = map[string][]storage.SeriesRef{}
-	p.m[allPostingsKey.Name][allPostingsKey.Value] = []storage.SeriesRef{1, 2, 3, 4, 6, 7, 8}
+	p.m[allPostingsKey.Name] = swiss.NewMap[string, []storage.SeriesRef](0)
+	p.m[allPostingsKey.Name].Put(allPostingsKey.Value, []storage.SeriesRef{1, 2, 3, 4, 6, 7, 8})
 
 	p.addFor(5, allPostingsKey)
 
-	require.Equal(t, []storage.SeriesRef{1, 2, 3, 4, 5, 6, 7, 8}, p.m[allPostingsKey.Name][allPostingsKey.Value])
+	srs, _ := p.m[allPostingsKey.Name].Get(allPostingsKey.Value)
+	require.Equal(t, []storage.SeriesRef{1, 2, 3, 4, 5, 6, 7, 8}, srs)
 }
 
 func TestMemPostings_ensureOrder(t *testing.T) {
 	p := NewUnorderedMemPostings()
-	p.m["a"] = map[string][]storage.SeriesRef{}
+	p.m["a"] = swiss.NewMap[string, []storage.SeriesRef](0)
 
 	for i := 0; i < 100; i++ {
 		l := make([]storage.SeriesRef, 100)
@@ -55,18 +57,19 @@ func TestMemPostings_ensureOrder(t *testing.T) {
 		}
 		v := strconv.Itoa(i)
 
-		p.m["a"][v] = l
+		p.m["a"].Put(v, l)
 	}
 
 	p.EnsureOrder(0)
 
 	for _, e := range p.m {
-		for _, l := range e {
+		e.Iter(func(_ string, l []storage.SeriesRef) bool {
 			ok := sort.SliceIsSorted(l, func(i, j int) bool {
 				return l[i] < l[j]
 			})
 			require.True(t, ok, "postings list %v is not sorted", l)
-		}
+			return false
+		})
 	}
 }
 
@@ -100,7 +103,7 @@ func BenchmarkMemPostings_ensureOrder(b *testing.B) {
 			// Generate postings.
 			for l := 0; l < testData.numLabels; l++ {
 				labelName := strconv.Itoa(l)
-				p.m[labelName] = map[string][]storage.SeriesRef{}
+				p.m[labelName] = swiss.NewMap[string, []storage.SeriesRef](0)
 
 				for v := 0; v < testData.numValuesPerLabel; v++ {
 					refs := make([]storage.SeriesRef, testData.numRefsPerValue)
@@ -109,7 +112,7 @@ func BenchmarkMemPostings_ensureOrder(b *testing.B) {
 					}
 
 					labelValue := strconv.Itoa(v)
-					p.m[labelName][labelValue] = refs
+					p.m[labelName].Put(labelValue, refs)
 				}
 			}