Use [DoltHub SwissMap](https://github.com/dolthub/swiss) instead of standard `map` in `tsdb/index.MemPostings`' table of series refs per label value, for performance.

#### Context

I've considered both the DoltHub and CockroachDB SwissMap implementations in the past, and found then that the former benchmarked better. Also, when @bboreham ran [prombench on DoltHub](https://github.com/prometheus/prometheus/pull/13644), no material difference could be observed, whereas [prombench for CockroachDB](https://github.com/prometheus/prometheus/pull/13646) at least appeared worse.

#### Benchmark results

<details>
<summary>PostingsForMatchers benchmark stats</summary>

```console
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
                                                                     │ postings-head-main.txt │      postings-head-dolthub.txt      │
                                                                     │         sec/op         │    sec/op     vs base               │
Querier/Head/PostingsForMatchers/n="1"-12                                        259.9n ±  3%   262.0n ±  2%        ~ (p=0.065 n=6)
Querier/Head/PostingsForMatchers/n="X"-12                                        257.6n ±  0%   259.8n ±  1%        ~ (p=0.058 n=6)
Querier/Head/PostingsForMatchers/n="1",j="foo"-12                                392.7n ±  0%   408.5n ±  1%   +4.02% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",j="foo"-12                                269.9n ±  1%   271.0n ±  1%        ~ (p=0.102 n=6)
Querier/Head/PostingsForMatchers/j="foo",n="1"-12                                393.4n ±  0%   406.4n ±  1%   +3.30% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",j!="foo"-12                               388.4n ±  1%   405.4n ±  1%   +4.39% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="2"-12                                 398.6n ±  1%   409.2n ±  1%   +2.65% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",j!="foo"-12                               271.7n ±  0%   270.7n ±  2%        ~ (p=1.000 n=6)
Querier/Head/PostingsForMatchers/i=~"1[0-9]",j=~"foo|bar"-12                     779.1n ±  0%   845.8n ±  1%   +8.57% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/j=~"foo|bar"-12                                 419.9n ±  1%   437.6n ±  1%   +4.21% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/j=~"XXX|YYY"-12                                 361.8n ±  1%   363.9n ±  1%        ~ (p=0.065 n=6)
Querier/Head/PostingsForMatchers/j=~"X.+"-12                                     327.7n ±  0%   303.1n ±  1%   -7.49% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~"(1|2|3|4|5|6|20|55)"-12                     651.3n ±  0%   697.6n ±  2%   +7.11% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i!~"(1|2|3|4|5|6|20|55)"-12                     713.9n ±  2%   767.2n ±  3%   +7.48% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~"X|Y|Z"-12                                   424.7n ±  0%   425.5n ±  1%        ~ (p=0.937 n=6)
Querier/Head/PostingsForMatchers/i!~"X|Y|Z"-12                                   476.0n ±  1%   493.0n ±  0%   +3.57% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~".*"-12                                      2.826m ±  1%   2.540m ±  2%  -10.12% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~"1.*"-12                                     3.858m ±  2%   3.495m ±  1%   -9.41% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~".*1"-12                                     2.453m ±  1%   2.178m ±  2%  -11.24% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~".+"-12                                      14.72m ±  2%   13.77m ±  2%   -6.50% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~".+",j=~"X.+"-12                             14.66m ±  1%   13.96m ±  2%   -4.81% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~""-12                                        14.63m ±  1%   12.74m ±  1%  -12.96% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i!=""-12                                        14.16m ±  2%   12.15m ±  1%  -14.21% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".*",j="foo"-12                        2.844m ±  1%   2.516m ± 21%        ~ (p=0.065 n=6)
Querier/Head/PostingsForMatchers/n="X",i=~".*",j="foo"-12                        299.2n ±  1%   299.6n ±  0%        ~ (p=0.738 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".*",i!="2",j="foo"-12                 2.858m ±  0%   2.558m ±  2%  -10.47% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!=""-12                                  14.28m ±  1%   12.27m ±  1%  -14.05% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j="foo"-12                          14.22m ±  1%   12.25m ±  1%  -13.81% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"X.+"-12                         14.15m ±  1%   12.14m ±  1%  -14.22% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"XXX|YYY"-12                     14.17m ±  1%   12.13m ±  1%  -14.38% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~"X|Y|Z",j="foo"-12                     573.3n ±  0%   582.6n ±  1%   +1.62% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!~"X|Y|Z",j="foo"-12                     677.9n ±  1%   701.0n ±  1%   +3.39% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",j="foo"-12                        14.73m ±  1%   13.81m ±  1%   -6.25% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~"1.+",j="foo"-12                       3.855m ± 14%   3.445m ±  1%  -10.65% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".*1.*",j="foo"-12                     10.84m ± 11%   10.59m ± 16%        ~ (p=0.589 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!="2",j="foo"-12                 14.89m ±  2%   14.48m ±  6%        ~ (p=0.240 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-12               19.40m ±  3%   18.20m ±  1%   -6.18% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-12             26.23m ±  3%   24.22m ±  1%   -7.64% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-12             345.9n ±  1%   347.0n ±  2%        ~ (p=0.145 n=6)
geomean                                                                          53.81µ         51.77µ         -3.79%

                                                                     │ postings-head-main.txt │      postings-head-dolthub.txt       │
                                                                     │          B/op          │     B/op      vs base                │
Querier/Head/PostingsForMatchers/n="1"-12                                          64.00 ± 0%     64.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X"-12                                          48.00 ± 0%     48.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",j="foo"-12                                  176.0 ± 0%     176.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",j="foo"-12                                  48.00 ± 0%     48.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j="foo",n="1"-12                                  176.0 ± 0%     176.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",j!="foo"-12                                 176.0 ± 0%     176.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!="2"-12                                   176.0 ± 0%     176.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",j!="foo"-12                                 48.00 ± 0%     48.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"1[0-9]",j=~"foo|bar"-12                       512.0 ± 0%     512.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j=~"foo|bar"-12                                   352.0 ± 0%     352.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j=~"XXX|YYY"-12                                   128.0 ± 0%     128.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j=~"X.+"-12                                       80.00 ± 0%     80.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"(1|2|3|4|5|6|20|55)"-12                       416.0 ± 0%     416.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i!~"(1|2|3|4|5|6|20|55)"-12                       480.0 ± 0%     480.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"X|Y|Z"-12                                     176.0 ± 0%     176.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i!~"X|Y|Z"-12                                     240.0 ± 0%     240.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".*"-12                                      1.531Mi ± 0%   1.531Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"1.*"-12                                     2.722Mi ± 0%   2.722Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".*1"-12                                     1.531Mi ± 0%   1.531Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".+"-12                                      12.22Mi ± 0%   12.22Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".+",j=~"X.+"-12                             12.22Mi ± 0%   12.22Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~""-12                                        12.22Mi ± 0%   12.22Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i!=""-12                                        12.22Mi ± 0%   12.22Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".*",j="foo"-12                        1.531Mi ± 0%   1.531Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",i=~".*",j="foo"-12                          48.00 ± 0%     48.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".*",i!="2",j="foo"-12                 1.532Mi ± 0%   1.532Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!=""-12                                  12.22Mi ± 0%   12.22Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!="",j="foo"-12                          12.22Mi ± 0%   12.22Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"X.+"-12                         12.22Mi ± 0%   12.22Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"XXX|YYY"-12                     12.22Mi ± 0%   12.22Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~"X|Y|Z",j="foo"-12                       240.0 ± 0%     240.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!~"X|Y|Z",j="foo"-12                       352.0 ± 0%     352.0 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",j="foo"-12                        12.22Mi ± 0%   12.22Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~"1.+",j="foo"-12                       2.722Mi ± 0%   2.722Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".*1.*",j="foo"-12                     5.906Mi ± 0%   5.906Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!="2",j="foo"-12                 12.22Mi ± 0%   12.22Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-12               14.95Mi ± 0%   14.95Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-12             18.14Mi ± 0%   18.14Mi ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-12               48.00 ± 0%     48.00 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                                          27.27Ki        27.27Ki       +0.00%
¹ all samples are equal

                                                                     │ postings-head-main.txt │      postings-head-dolthub.txt      │
                                                                     │       allocs/op        │  allocs/op   vs base                │
Querier/Head/PostingsForMatchers/n="1"-12                                          3.000 ± 0%    3.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X"-12                                          3.000 ± 0%    3.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",j="foo"-12                                  7.000 ± 0%    7.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",j="foo"-12                                  3.000 ± 0%    3.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j="foo",n="1"-12                                  7.000 ± 0%    7.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",j!="foo"-12                                 7.000 ± 0%    7.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!="2"-12                                   7.000 ± 0%    7.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",j!="foo"-12                                 3.000 ± 0%    3.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"1[0-9]",j=~"foo|bar"-12                       14.00 ± 0%    14.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j=~"foo|bar"-12                                   8.000 ± 0%    8.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j=~"XXX|YYY"-12                                   6.000 ± 0%    6.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j=~"X.+"-12                                       4.000 ± 0%    4.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"(1|2|3|4|5|6|20|55)"-12                       12.00 ± 0%    12.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i!~"(1|2|3|4|5|6|20|55)"-12                       15.00 ± 0%    15.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"X|Y|Z"-12                                     7.000 ± 0%    7.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i!~"X|Y|Z"-12                                     10.00 ± 0%    10.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".*"-12                                        6.000 ± 0%    6.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"1.*"-12                                      11.12k ± 0%   11.12k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".*1"-12                                       4.000 ± 0%    4.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".+"-12                                       100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".+",j=~"X.+"-12                              100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~""-12                                         100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i!=""-12                                         100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".*",j="foo"-12                          10.00 ± 0%    10.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",i=~".*",j="foo"-12                          3.000 ± 0%    3.000 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".*",i!="2",j="foo"-12                   14.00 ± 0%    14.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!=""-12                                   100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!="",j="foo"-12                           100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"X.+"-12                          100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"XXX|YYY"-12                      100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~"X|Y|Z",j="foo"-12                       10.00 ± 0%    10.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!~"X|Y|Z",j="foo"-12                       14.00 ± 0%    14.00 ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",j="foo"-12                         100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~"1.+",j="foo"-12                        11.12k ± 0%   11.12k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".*1.*",j="foo"-12                      40.97k ± 0%   40.97k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!="2",j="foo"-12                  100.0k ± 0%   100.0k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-12                111.2k ± 0%   111.2k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-12              141.1k ± 0%   141.1k ± 0%       ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-12               3.000 ± 0%    3.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                                            233.3         233.3       +0.00%
¹ all samples are equal
```
</details>