@codesome 2024-06-01
@aknuds1 swissmap seems to be slightly worse in query engine timings (inner_eval), slightly worse in CPU at times, and no difference in memory.