@codesome 2024-06-01
Is this PR any different from https://github.com/prometheus/prometheus/pull/13644?