@aknuds1 2024-06-14
> swissmap seems to be slightly worse in query engine timings (inner_eval), slightly worse in CPU at times, and no difference in memory.
> Is this PR any different from https://github.com/prometheus/prometheus/pull/13644?

@codesome Essentially it's the same PR, but #13644 was opened just to run prombench and compare. Bryan [couldn't observe any material difference then](https://github.com/prometheus/prometheus/pull/13644#issuecomment-1964303842).

Since Bryan couldn't see any difference then, do you think the differences you could spot are possibly by chance? Do I recall correctly in prombench not testing regular expressions? All the benchmark improvements are with regular expression matchers (only CPU, memory remains the same after I merged in the latest changes from main).